requirements.txt
Django==3.0.6
django-braces==1.14.0
django-oauth-toolkit==1.3.2
django-simple-history==2.10.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2-binary==2.8.5
